+++ R/map_layer_screengrid.R
@@ -41,7 +41,7 @@ mapdeckScreengridDependency <- function() {
 #' ))
 #'
 #' df <- df[ !is.na(df$lng), ]
-#' df$weight <- sample(1:10, size = nrow(df), replace = T)
+#' df$weight <- sample(1:10, size = nrow(df), replace = TRUE)
 #'
 #' mapdeck( style = mapdeck_style('dark'), pitch = 45 ) %>%
 #' add_screengrid(