+++ inst/htmlwidgets/lib/tile3d/tile3d.js
@@ -1,11 +1,11 @@
 
-function add_cesium( map_id, map_type, tile_data, layer_id, ion_token ) {
+function add_cesium( map_id, map_type, tile_data, point_size, layer_id, ion_token ) {
 
 	const tileLayer = new deck.Tile3DLayer({
 		map_id: map_id,
 		id: 'cesium-'+layer_id,
 		data: tile_data,
-		pointSize: 2,
+		pointSize: point_size,
 		loader: loaders.CesiumIonLoader,
 		loadOptions: {
 			tileset: {
@@ -33,7 +33,6 @@ function add_cesium( map_id, map_type, tile_data, layer_id, ion_token ) {
 
   */
   //md_layer_view( map_id, map_type, layer_id, focus_layer, bbox, update_view );
-
 }
 
 
@@ -43,7 +42,6 @@ function add_i3s( map_id, map_type, tile_data, layer_id ) {
 		map_id: map_id,
 		id: 'i3s-'+layer_id,
 		data: tile_data,
-		//pointSize: 2,
 		loader: loaders.I3SLoader
 	});