+++ NEWS.md
@@ -1,6 +1,7 @@
 # v0.3.4
 
 * add_cesium and add_i3s layers
+* add Linking to geometries ready for migration
 
 # v0.3.3